Make urubu smarter about ignoring things.
If a directory is ignored, it should not be looked in forr
content. For instance, if I create a virtual environment in
_venv that has urubu installed in it, it should not be traversed
for markdown files (which currently causes errors due to unfound
references).

There are also times we need files for building that we don't
want to be part of the output. One example is requirements.txt
for pip. Rather than having a hard-coded list of every possible
file, I've added a list to _site.yml called "ignore_patterns".
This can be a list of file names or globs that will not be
processed or copied.