add baseurl option to site configuration
This is meant to mirror the baseurl option in Jekyll:
http://jekyllrb.com/docs/configuration/#serve-command-options
It allows you to specify a prefix for all local URLs generated
within your site. This is necessary when your site will be served
from a URL that has more than just the hostname. For example,
on GitHub Pages sites are served from
http://username.github.io/project_name/, so Urubu needs to include
that /project_name/ in generated URLs pointing to local content.

The baseurl should be specified in _site.yml with no beginning
or trailing slashes, e.g.:

    baseurl: prefix