Add option for jinja2 to fail on undefined vars
If a variable is missing from a template, the current behaviour of
urubu is to treat it as an empty string. So if you forget to define (or
define it wrong), the resulting HTML will be empty and you might not
realise it.

This adds an optional argument to `_site.yml` called `strict_undefined`.
If omitted or `false`, urubu behaves normally. If defined and `true`,
then will add `undefined=jinja2.StrictUndefined` to `jinja2.Environment`
and jinja will fail with an error when an undefined variable is
encountered.